requirements.txt
streamlit
python-dotenv
langchain
langchain-community
langchain-core
langchainhub
langchain-huggingface
langchain-groq
transformers
sentence-transformers
